requirements.txt
# Core dependencies
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
tqdm==4.66.1
lxml==4.9.3

# Web scraping - by preference order
crawl4ai==0.3.2
pyppeteer==1.0.2
selenium==4.15.2
beautifulsoup4==4.12.2
webdriver-manager==4.0.1
scrapinggraph-ai==0.1.5

# Data processing and visualization
matplotlib==3.8.2
scikit-learn==1.3.2

# Other dependencies
g4f==0.5.0.1
importlib-metadata==8.6.1
litellm==1.65.0
openai==1.69.0
pip==25.1.1
zipp==3.21.0
